+++ dist_apps/cc_pull/gen_sync.hh
@@ -1,6 +1,6 @@
 #include "galois/runtime/SyncStructures.h"
 
-GALOIS_SYNC_STRUCTURE_BROADCAST(comp_current, unsigned long long);
-GALOIS_SYNC_STRUCTURE_REDUCE_SET(comp_current, unsigned long long);
-GALOIS_SYNC_STRUCTURE_REDUCE_MIN(comp_current, unsigned long long);
+GALOIS_SYNC_STRUCTURE_BROADCAST(comp_current, uint32_t);
+GALOIS_SYNC_STRUCTURE_REDUCE_SET(comp_current, uint32_t);
+GALOIS_SYNC_STRUCTURE_REDUCE_MIN(comp_current, uint32_t);
 GALOIS_SYNC_STRUCTURE_BITSET(comp_current);
